+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.21.34
+
+- Export missing `PanaProcessResult`.
+
 ## 0.21.33
 
 - Do not block repository check when `pubspec.yaml` is in path that is not accepted.